Sound player: No longer throws an error due to an empty JSON object. No longer tries to download before the auth is set. Sets robot upon change of robot, not just the original on load.
Auth:
Catches bad auth and disconnects.
Has a callback for an alert error...needs to actually do something the user can see (such as a popup...which is not trivial or else it would be done).
Callback is in gui.js:91

Frontpage:
Changed scaling to a static size on frontend main site.